+++ R/metabin.R
@@ -670,7 +670,7 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
                     level.hetstat = gs("level.hetstat"),
                     tau.preset = NULL, TE.tau = NULL,
                     tau.common = gs("tau.common"),
-                    detail.tau = "",
+                    detail.tau = NULL,
                     #
                     method.I2 = gs("method.I2"),
                     #