+++ R/metacor.R
@@ -278,7 +278,7 @@ metacor <- function(cor, n, studlab,
                     level.hetstat = gs("level.hetstat"),
                     tau.preset = NULL, TE.tau = NULL,
                     tau.common = gs("tau.common"),
-                    detail.tau = "",
+                    detail.tau = NULL,
                     #
                     method.I2 = gs("method.I2"),
                     #