+++ R/metaprop.R
@@ -608,7 +608,7 @@ metaprop <- function(event, n, studlab,
                      level.hetstat = gs("level.hetstat"),
                      tau.preset = NULL, TE.tau = NULL,
                      tau.common = gs("tau.common"),
-                     detail.tau = "",
+                     detail.tau = NULL,
                      #
                      method.I2 = gs("method.I2"),
                      #